updated balance endpoint scenario for INACTIVE status and split happypaths to new spec files
